+++ web/src/cms/types.ts
@@ -12,7 +12,6 @@ export {
   DEFAULT_LOCALE,
   LOCALE_URL_PREFIX,
   MULTIPLE_LOCALES,
-  SITE_NAME,
 } from '@astroload/cms/src/site-config'
 export { stripLocalePath, stripLocalePathsDeep } from '@astroload/cms/src/stripLocalePath'
 export type PageData = Page | Post | Author